requirements.txt
requests==2.32.3
python-dotenv==1.0.1
rich==13.9.4
click==8.1.8
packaging==24.2
tomli==2.2.1; python_version < "3.11"
beautifulsoup4==4.12.3
pyyaml==6.0.2
